+++ changelog.md
@@ -318,6 +318,9 @@
 # 7.3.3
 
 * Parents with same gender are now properly registered
+
+# 7.3.4
+
 * Engagement rings now set the relationship to engaged
     * Engaged villagers won't marry someone else
 * Gifting a bouquet prevents villagers from marrying other villagers